training/pyproject.toml
[tool.poetry]
name = "training"
version = "0.1.0"
description = ""
authors = ["Daniel Wu <dwu359@gatech.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9, <3.10"
boto3 = "^1.28.15"
flask = "^2.3.2"
django = "^4.2.3"
django-ninja = "^0.22.2"
scikit-learn = "^1.3.0"
pandas = "^2.0.3"
torch = "^2.0.1"
torchvision = "^0.15.2"
firebase-admin = "^6.2.0"
pre-commit = "^3.3.3"
click = "^8.1.7"
requests = "^2.31.0"
django-cors-headers = "^4.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
ggshield = "^1.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
